chore: convert feature flag tests to unit tests
